requirements.txt
aiohttp==3.8.4
aioice==0.9.0
aiortc==1.6.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
distro==1.9.0
dnspython==2.6.1
frozenlist==1.4.1
google-crc32c==1.5.0
gradio==3.26
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
multidict==6.0.5
numpy==1.24.3
openai==1.12.0
openai-whisper
opencv-python==4.8.1.78
pillow==10.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.1.0
pylibsrtp==0.10.0
pyOpenSSL==24.0.0
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
torch
git+https://github.com/openai/whisper.git
